refactor(swc/plugin): align exports
